[Storage] Move some truncation helper functions to its own file.
