+++ storage/aptosdb/src/utils/mod.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod iterators;
+pub(crate) mod truncation_helper;
